fix partition reporting
